perf(vram): cut the resident footprint of a loaded checkpoint
Reported on #551 from an RTX 5060 Ti: large-v3-turbo in FP16 builds fine
now, but holds roughly 3-4x its FP16 weight size at runtime, which makes it
impractical next to any other GPU tenant on a 16 GB card. --max-workspace-mb
does not help — it bounds tactic search during the build, not what a loaded
engine reserves. Three things that do:

Shared engine scratch. Every engine in a checkpoint now draws its layer
scratch from one pool sized to the largest of them, instead of each context
reserving a private pool. A KV-mode checkpoint is four engines (encoder plus
cross-K/V, prefill and step) that run strictly one at a time — encode, then
the decode loop — which is the condition sharing requires. Falls back to
private pools on a torch2trt without SharedDeviceMemory rather than failing
to load.

One serialized plan live at a time. load() held every engine's serialized
bytes for the whole call while deserialize_cuda_engine made its own copy of
each, roughly doubling the load-time peak. Each plan is now popped and
dropped as it is consumed, the checkpoint is restored on the host rather
than the GPU, and a build that runs in this process reclaims its ONNX and
builder residue before the engine loads on top of it — the reported 7324 MiB
post-build figure against a lower figure on a fresh load is that residue.

FP16 tied embedding. The token embedding is the largest thing torch itself
keeps resident (~265 MiB in FP32 for the multilingual vocab) and it feeds
decoder engines that are FP16 either way, so it is kept in FP16 when the
decoder is. Consumers cast the hidden state to the weight dtype, so this is
dtype-safe in both modes and logits are still returned as FP32.

The first two are pure footprint. The third changes the precision of the
final logits projection, so it wants a WER check on real hardware before
this merges.

+++ whisper_trt/_decoder.py
@@ -277,7 +277,7 @@ class TextDecoderTRTKV(nn.Module):
         last_hidden, self_kv = self.engines.prefill(hidden_in, cross_kv, mask)
         last_hidden = self.ln(last_hidden)
         weight = self.token_embedding.weight.to(device)
-        logits = (last_hidden @ torch.transpose(weight, 0, 1)).float()
+        logits = (last_hidden.to(weight.dtype) @ torch.transpose(weight, 0, 1)).float()
         return logits, self_kv
 
     @torch.no_grad()
@@ -296,7 +296,7 @@ class TextDecoderTRTKV(nn.Module):
         hidden, new_self_kv = self.engines.step(hidden_in, self_kv, cross_kv)
         hidden = self.ln(hidden)[:, -1:, :]
         weight = self.token_embedding.weight.to(device)
-        logits = (hidden @ torch.transpose(weight, 0, 1)).float()
+        logits = (hidden.to(weight.dtype) @ torch.transpose(weight, 0, 1)).float()
         return logits, new_self_kv
 
     def forward(self, *args: Any, **kwargs: Any) -> Any:
@@ -361,7 +361,7 @@ class TextDecoderTRT(nn.Module):
         hidden = self.engine(hidden, xa, cast(Tensor, self.mask).to(xa.device))
         hidden = self.ln(hidden)[:, -1:, :]
         weight = self.token_embedding.weight.to(hidden.device)
-        return (hidden @ torch.transpose(weight, 0, 1)).float()
+        return (hidden.to(weight.dtype) @ torch.transpose(weight, 0, 1)).float()
 
     def summary(self) -> str:
         """Return a short human-readable component summary."""
